requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
pdfplumber
python-docx
